Enable different coupled settings for ensemble (#2030)
Adds the ability to turn off some coupled components for the ensemble.

To enable this, many settings are moved out of `config.fcst` and into `config.ufs`
so they are calculated after any overrides of `DO_*` have been made. Some
settings also had to be moved out of `config.base` for the same reason. A
switch to turn on gocart is added to `config.ufs`, but unlike the other switches
this one takes no arguments.

The name of the ufs configure template used is converted from a partial file
name to the full file name.

Resolves #1692