Add wave init to GEFS
Adds the wave init job to GEFS. ww3_multi and ww3_shell templates are
copied from the GFS versions.

The GEFS restart frequency is changed to not be zero, and the comment
about not setting it to zero is copied from the gfs config.base. The
restart interval for the GEFS perturbed members is changed to equal that
of the control.

The `config.wave` file is changed from a link to the GFS version to a
seperate copy so it can be modified to account for the different `RUN`.

Also fixes a bug where `RUNwave` was not defined for ensemble forecasts.

Resolves #2017