Make proper j-job for MET+
Creates a proper j-job for the metp job, moving much of the material
from the `jobs/rocoto` script there (some of it within `jjob_header`).

Also updates the old `COMIN` variable for the COM refactor (although
it does not appear used in the verify scripts currently).

This is a partial step towards restoring/improving metp functionality
and remains untested.

Refs #1575