Add build options to control build GSI and UFSDA
The move to submodules eliminates the capability to build based on
the existance of directories, since they will now always be checked
out. To avoid unnecessary compilation time, the build script is
updated to take the same options as the old checkout script to build
GSI (-g) and UFSDA (-u). The default is to not build these.