Add basic atmos products for GEFS and refactor resources (#2216)
* Add basic atmos products for GEFS

Adds basic atmosphere products for GEFS members (no mean or spread).

To facilitate the addition of GEFS parameter lists, the existing GFS
lists are renamed. They are also relocated to `parm/products/` direc
tory instead of `parm/post/`. GEFS v12 uses different parameter lists
for different resolutions, but for this implementation, the 0p25 lists
are used for all resolutions. However, to hedge against the final con-
figuration, all of the GEFS parameter lists have been added to parm.
Implementation of different parameter lists for each resolution will
take additional refactoring of the atmos_products job, as it currently
assumes all resolutions use the same lists.

The new GEFS rocoto task is implented as two loops of metatasks (made
possible by PR #2189). The member is set at the rocoto level, so the
setting in `config.base` had to be updated to accept a pass-through.
The generation of a forecast hour list is moved up to its own function
in `rocoto.tasks`, separate from the post groups which are not used.
This can be used as a model for future metatasks (or refactoring of old
ones). `rocoto_viewer.py` does function with the nested metatasks,
though only one level of folding is available and the metatask name
is that of the lower metatask (though it encompasses the entire outer
metatask).

Additionally, the GEFS `config.resources` had diverged from the GFS
version, which included not having a section for the atm products job.
This led to the refactoring of the GFS `config.resources`, including
switching to using a `case` statement instead of sequential `elif`.
This refactored script was then copied over to GEFS, replacing the old
one, with all jobs not currently implemented for GEFS removed (except
wave post, whose inclusion is immenent). New blocks can be copied over
from GFS as they are added.

Resolves #823