Add wave post jobs to GEFS
Adds the wave post jobs for gridded and points to GEFS. Boundary
point jobs are added even though the current GEFS buoy file does
not contain any.

To facilitate the dependencies and make things consistent, the
forecast job is altered to all be a metatask of fcst instead of
mimicking EnKF using efcs. This cleans up the dependency.

Resolves #827