Add `POSTAMBLE_CMD` into preamble postamble (#2235)
This PR adds the ability to run a command at the end of jobs via the
preamble's postamble function. A new command can be set via
`POSTAMBLE_CMD` and will be invoked at the end of jobs. Users can
add the command to the top of an env file to have every job run it or
it can be placed within a job if-block in the env file to run for just that
job.

Resolves #1145