GOCART Emission updates for GEFSv13 and GDAS  (#2201)
This PR addresses several things needed for more recent simulations using
GOCART2G as well as preparing for the GEFSv13 30 year run.  The main
updates are:

- Update CEDS to use monthly emission files instead of daily.  This will
drastically reduce the number of files needed. No science change
- Update biogenic MEGAN inputs to use a climatology rather than the
offline biogenic previously used.  This is needed to deal with simulations
where the current dataset is not available. 
- Update volcanic SO4 emissions to use degassing emissions only.  This is
to support the need for more recent simulations where data is not available.