Update typing hint for WCOSS version of python (#2238)
The typing hint `typing.List` was deprecated with python 3.9 in favor
of using the primitive `list[str]`, but the functional version of python
on WCOSS2 is <3.9, causing `setup_xml.py` to fail there. This replaces
`list[str]` as a typing hint with the deprecated form until the supported
version on WCOSS2 is >=3.9.