STY: Use seq to generate forecast hour list instead of for loops
I'm running a year-long forecast, which means I get a large portion of the log file dedicated to these loops.  

`seq ${START} ${STEP} ${STOP}` will generate a sequence going from START to STOP by STEP, including STOP if relevant.  That seems to be the purpose of these loops.

I would like this to be two lines in the log, rather than a few hundred, and this may also be faster, though probably more for reasons of fewer writes to disk than because bash isn't designed for arithmetic.