Retire cycle-specific FHMAX_GFS variables (#2278)
This PR retires the `FHMAX_GFS_${cyc}` variables that allowed users to specify
different gfs forecast lengths for each cycle. This function is no longer supported
in global-workflow. The `FHMAX_GFS_*` variables will be removed and will no
longer be checked to set the final `FHMAX_GFS` variable. The same forecast
length will be set for every cycle.

This PR also includes a small fix to add new post parm files into the `.gitignore`
file. This was intended to be included in a different PR but that PR is on hold for
further testing so it is being included here to get it into `develop` sooner.

Resolves #2218