Update GSI hashes to 2024 Feb 13
Updates the hashes for gsi_enkf and gsi_monitor to pick up changes
to the detect-machine scripts. This should eliminate the build
interference/race condition that was causing CI tests to fail.

This is a bit of a paste-over of the problem. A more thorough
solution should be implemented later so that individual build scripts
do not interact when running build_all.