Add unstructured grid for HR3/GFS (#2230)
This adds the capability to use unstructured grids in the global workflow,
which will be used in HR3.  There are new fix files for a low-resolution
100km grid and a grid closer to our targeted GFSv17 grid which has the
resolution combined from the older multi_1 and GFSv16 grids.

The fix file update is here: NOAA-EMC/global-workflow#2229 

Note: This now means that GFS tests need a new build option:  
`./build_all.sh -w` So that PDLIB=ON is turned on for compiling relevant
UFS and WW3 codes. 

Resolves NOAA-EMC/global-workflow#1547