Fix rocoto forecast hour determination for GEFS
The function that generates the list of forecast hours for rocoto
was trying to use variables that are not defined for GEFS, causing
workflow generation to fail. The function is updated to not try to
load these variables before loading the ones actually used for GFS/
GEFS.

Also turns GEFS CI test back on.