Update stage IC to handle ocean perturbations (#2364)
This PR adds the option to stage ocean perturbation files for ensemble
forecasts. These perturbation files are used in GEFS forecasts.
A new variable is introduced in config.base to use the ocean
perturbation files.
This PR does not include using these perturbation files. A future PR
will address this.