Allow non-matching bufr table
PR #2383 added a new bufr table, along with a fatal error if there was
not one that matched the resolution. However, workflow usually uses a
low resolution to test, and it works fine even if the bufr parm file is
a mismatch, so turned the fatal error into a warning and links the bufr
parm file as before.