Add initial GSI-based soil analysis capability (#2263)
First set of changes for adding the new soil analysis, from the
assimilation of screen-level T and q. The changes here enable the
screen-level observations to be assimilated in the Hybrid (Var and EnKF)
update, and the soil temperature and soil moisture updates to be made in
the EnKF only. The functionality is turned on by setting GSI_SOILANAL to
YES in config.base.

Resolves #1479