Modify APP based on RUN
There is a need to change which coupled components are on depending
on the current `RUN`. To facilitate this, the `APP` is modified prior
to the setting of the `DO_<component>` variables based on `RUN`,
turning off components as desired.

This new system also replaces the `DO_<component>_ENKF` switches that
were formerly used to turn components off for the ensemble.

Resolves #2318