Uploading error logs to GitHub from Jenkins CI Runs (#2429)
This PR enhances the user experience within GitHub for when errors occur
during build and running of CI cases from within the PR messages. This
is done by uploading the error logs to GitHub Gists and then publishing
the links to them along with the full paths of the logs on disk.

This PR adds the python Class **GitHubPR** in
`${HOMEgfs}/ci/scripts/utils/githubpr.py` by inhearting the GitHub Class
from **pyGitHub**. We use this module to introduce a helper python
utility that can publish a list of log files into a GitHub Gist and/or
the designated branch **error_logs** in the **emcbot** repo
**ci-global-workflows** for storing error log files for review from any
git configured terminal. This upload feature will also create
persistence of errors over time.

Also the `build_all.sh -k` script has been updated to support a "quick
kill" feature (thanks David) that stops the parallel builds whenever one
fails and creates an error_log file that has the paths to the error
files that are also uploaded and published in the PR messages in GitHub.

Co-authored-by: TerrenceMcGuinness-NOAA <terry.mcguinness@noaa.gov>
Co-authored-by: terrance.mcguinness <terrance.mcguinness@noaa.gov>
Co-authored-by: DavidHuber <david.huber@noaa.gov>
Co-authored-by: Walter Kolczynski - NOAA <Walter.Kolczynski@noaa.gov>