Fix rotating member bugs (#2443)
When PR #2427 introduced the rotating subset of member guess states for
the early-cycle EnKF, the rotating member calculation function was
omitted from the DOSFCANL_ENKF if block in the enkf surface script. This
PR adds this feature to this section. This PR also removes hard coded
values that were included in this function and are replaced with a
variable equal to the number of late cycle members.

Resolves #2441