Flip the build for GFS and GEFS with waves (#2462)
This PR:
- builds the unstructured waves as default when invoked with
`build_all.sh`.
- updates the CI testing infrastructure to handle this change.
- updates the documentation to reflect this change in behaviour.

To build with structured grid for waves, the flag `-w` needs to be
passed to `./build_all.sh`. Previously, this would trigger the
unstructured grid. Since GFSv17 will move to unstructured grid as
default for waves, this makes it easier for a majority of developers.

Resolves #2461