Update fbwind for COM refactor (#2479)
Updates fbwind job for the COM refactor and some other cleanup.

This works on WCOSS but not on Orion. There seems to be a problem with
either `grbindex` or `GETGB()` on Orion that causes the executable to be
unable to read the grib1 index file. The grib1 data file produced there
seems fine. Haven't checked Hera yet, maybe there is a problem with the
spack-stack build of `grbindex`.

Resolves: #2160
Refs: #289