Add mean/spread for atmos grib2 (#2482)
Adds the basic mean and spread grib2 products for the atmosphere.

The new mean/spread files are named according to the new convention,
though `pres_${grid_type}` may be changed to `pgrb_${grid_type}`. Either
way, the input filenames will need to be updated when that change is
made. For now, the mismatch means the unlettered grib files result in
`pres_`.

Resolves #2296
Refs #1522