Add restart on failure capability for the forecast executable (#2510)
This PR:
- enables restart capability of the forecast executable from a previous
failure.
- saves restarts during the run in a new `DATA` structure. The current
`DATA` structure:
![current
`DATA`](https://github.com/NOAA-EMC/global-workflow/assets/11394126/03383e2f-b7f8-43e0-8b78-c8f37a79ab84)
is being replaced by:
![Screenshot 2024-04-19 at 12 55
44 PM](https://github.com/NOAA-EMC/global-workflow/assets/11394126/8ab6e6df-bbdb-43cf-b0dc-8e066f537ee7)
where, the colored boxes are described as:
![Screenshot 2024-04-19 at 12 56
14 PM](https://github.com/NOAA-EMC/global-workflow/assets/11394126/30b20e50-6cc8-4433-988a-02d5b484e7b5)
- saves model output from `MOM6` and `CICE` within `MOM6_OUTPUT/` and
`CICE_OUTPUT/` sub-directories. This is done to keep the run directory
clean and easily identify component output.

This PR also:
- replaces link with copy. This enables the creation of a `DATA`
directory that is self-contained and can be used to diagnose issues
during failures. This is a NCO EE2 requirement and addresses part of an
outstanding bugzilla.

In the process of enabling the restart capability, functionality from
`forecast_postdet.sh` is moved to `forecast_predet.sh` that does not
depend on the outcome of `forecast_det.sh`. `forecast_det.sh` determines
where the initial conditions will come from; `COM` in the case of a
clean run or `DATArestart` in the case of a `RERUN`.
This should make it easier to separate **static** configuration and data
(fix files, etc) from **runtime** configuration (namelists, etc) and
data (initial conditions)

Additionally, this PR:
- adds 3 utility shell scripts in `test/`.  
  - 'nccmp.sh` - compare netCDF files using `nccmp`
  - `g2cmp.sh` - compare grib2 files using `wgrib2`
- `f90nmlcmp.sh` - compare Fortran90 nml files using `f90nml` (Requires
modulefiles to load `py-f90nml` module on RDHPCS platforms)
They are not used in the workflow, but are useful for users to compare
files.

Resolves #2273

Co-authored-by: Walter Kolczynski - NOAA <Walter.Kolczynski@noaa.gov>