Hotfix to disable STALLED in CI as an error (#2523)
This **_hotfix_** to the last CI updates disables the STALLED feature by
not flagging it as an error until a more succinct algorithm is
determined (i.e. the STALLED state will be at par with RUNNING).

A second minor bug is in the event that no error log is produced when a
Case fails, the error of the fail is still reported to the general user
in GitHub. (Jenkins enabled users can always see the full comprehensive
state of system by following the links at the bottom of the PR).