Remove implicit symlink names
Lustre has a defect under Rocky 9 that results in symlink sometimes
failing when the link name is not explicit. This updates all link
creation to use explicit names.

`config.base` is updated to turn off two monitor jobs on Hercules
because the executables are not yet built there. This, combined with
the previous change, should make workflow available for use on
Hercules.

Also removes the redundant utility names for NCP, NLN, etc. in the
gdas scripts that are already defined in `config.base`.

Resolves #2131
Resolves #2522