Add extended product case for WCOSS
Adds a new version of the atm3DVar test that runs the full forecast
length and produces most of the secondary products. For now, this
test will only run on WCOSS due to gempak failures on other machines
as well as computational needs. On other machines, the original
version will run (the original version will not run on WCOSS).

Resolves #2132