Updated GEMPAK version and APRUN launcher. (#2555)
This PR addresses issue #2248 and #2513. The following is accomplished:

- Updates the parallel executable for GEMPAK applications;
- Updates the GEMPAK version for RDHPCS Hera.

Describe your changes. Focus on the *what* and *why*. The *how* will be
evident from the changes. In particular, be sure to note any interface
changes, such as command line syntax, that will need to be communicated
to users.

At the end of your description, please be sure to add the issue this PR
solves using the word "Resolves". If there are any issues that are
related but not yet resolved (including in other repos), you may use
"Refs".

Resolves #2248 
Resolves #2513 

Co-authored-by: David Huber <69919478+DavidHuber-NOAA@users.noreply.github.com>
Co-authored-by: henrywinterbottom-wxdev <henry.winterbottom.wxdev@gmail.com>