Add CI test for products (#2567)
Adds a new version of the atm3DVar test that runs the full forecast
length and produces most of the secondary products. For now, this test
will only run on WCOSS due to gempak failures on other machines as well
as computational needs. On other machines, the original version will run
(the original version will not run on WCOSS). AWIPS remains off for now
in this extended test due to a bug involving tocgrib2 and the convective
precip fields (see #2566).

The new test runs for 4½ cycles and the full 384-hr forecast length to
ensure all gempak scripts are exercised. Since the cycle throttle is 3
and the bulk of the time is in the free forecast, the cycles run mostly
concurrently so it doesn't extend the total test time too much beyond
that of a single 384-hr forecast.

Fixes a bug in NPOESS that was introduced when the post filenames were
reverted to the previous format for the GOES products until the final
filenames are determined (#2499).

Also removes the AWIPS g2 job from the rocoto mesh to complete the
retirement of grib1 products.

Resolves #2132
Resolves #2445