Updates for cold start half cycle, then continuing with IAU for WCDA (#2560)
This PR allows us to run C384 S2S with IAU, but starting with the first
half-cycle as a cold-start. This will be necessary for cycled testing as
we build towards the full system for GFSv17. This updates the copying of
the restarts for RUN=gdas for both ocean and ice copying what the atm
model is doing. It also reduced the amount of restart files from 4 to 3.

Other updates: 
* Add DOJEDI ocean triggers for archiving certain files update from:
@CatherineThomas-NOAA
* Adds COPY_FINAL_RESTARTS option to turn on/off copying the last
restart file to COM. Defaults to off...
* Defines model_start_date_current_cycle & model_start_date_next_cycle
to help with knowing which IC to grab.

Refs #2546

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>