Remove if statement that removes zeroes from fhrs
An if statement has been removed from fhrs in tasks.py. The purpose of this if statement was to remove the 0 lead hour for ocean and ice components. Since this function is already performed in gefs_tasks.py, this if statement is not needed. Furthermore, keeping this if statement caused an "x not in list" error.