Switch to Rocky 9 built external packages on Hercules (#2608)
The workflow was updated to use modules built on Rocky 9, but the
external packages (like prepobs) were still pointing to the versions
built on CentOS (Orion). This transitions to packages built on Rocky 9.

Updating of the tracker package has been deferred until later. As such,
the tracker jobs have been disabled by returning immediately if they are
on Hercules. Since these jobs are small, resource-wise, it should not
meaningfully impact turnover time.