Parameterize some things in config.atmanl and config.atmensanl (#2661)
This PR adds some parameters in config.atmanl and config.atmensanl that
can be altered with the defaults.yaml. The motivation is to make these
files match those in the GDASApp JJOB tests (example:
https://github.com/NOAA-EMC/GDASApp/blob/develop/test/atm/global-workflow/config.atmanl),
so we can just use the Global Workflow config.atmanl and
config.atmensanl in the tests rather than custom ones in GDASApp that
have to be separately updated every time the ones in the Global Workflow
are updated.