Add task to process reforecast variables to save on WCOSS2 (#2680)
# Description
This PR adds an optional task to the global-workflow to process a subset
of ocean, ice, wave and atmosphere products to be saved on WCOSS2 for
the GEFSv13 reforecast. This task is designed to process GEFS variables
so that specific reforecast product requirements are met. A new variable
in `config.base` called `DO_EXTRACTVARS` enables this task, which is
currently called `extractvars`. `DO_EXTRACTVARS` is set to `NO` by
default and is specifically a task designed to be executed for the
GEFSv13 reforecast.

Refs #1878 

# Type of change
- New feature (adds functionality)

# Change characteristics
- Is this a breaking change (a change in existing functionality)? NO
- Does this change require a documentation update? NO

# How has this been tested?
This has been cloned and tested on WCOSS2. This will need to be tested
on Hera and other platforms on which the reforecast may be running.

# Checklist
- [ ] Any dependent changes have been merged and published
- [x] My code follows the style guidelines of this project
- [x] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [x] My changes generate no new warnings
- [x] New and existing tests pass with my changes
- [ ] I have made corresponding changes to the documentation if
necessary

---------

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>