Remove f000 from atmos rocoto tasks for replay cases (#2778)
The main purpose of this PR is to remove the f000 from
atmosphere-related rocoto tasks when `REPLAY_ICS` is set to `YES`. In
cases where `REPLAY_ICS` is `YES` and `OFFSET_START_HOUR `is greater
than `0`, it becomes necessary to have the first lead hour set to
`OFFSET_START_HOUR ` for the atmosphere-related rocoto tasks. For
example, when `OFFSET_START_HOUR ` is set to `3`, then the minimum lead
time for the atmos_prod and atmos_ensstat rocoto tasks needs to be 3 and
the minimum lead time for the ocean_prod rocoto task needs to be 6
(assuming `FHOUT_OCN` is 6). This PR makes this rocoto workflow set up
possible by removing 0 from fhrs for atmosphere-related tasks in
`gefs_tasks.py` when replaying.

This PR also moves where f000 is being removed for the ocean_prod and
ice_prod tasks. The if-block that performs this f000 removal has been
moved from `tasks.py` to `gefs_tasks.py` and `gfs_tasks.py`.

---------

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>