Change fcst segments to breakpoints, move to config.fcst
Changes the way forecast segments are defined. Restores the original
`FHMIN_GFS` and `FHMAX_GFS` and then adds a local `breakpnts` variable
that contains the intermediate stopping points (if any). The original
list of segment endpoints is then constructed from that.

The determination of the `FHMIN` and `FHMAX` based on the segment is
moved from `config.base` to `config.fcst`. This required adding some
additional checks in `config.fcst` to clip other `FHMAX` variables to
`FHMAX`.