Add fixes to products for when REPLAY IC's are used  (#2755)
This PR fixes a couple issues that arise when replay initial conditions
are used. These issues only occur when `REPLAY_ICS` is set to `YES` and
`OFFSET_START_HOUR` is greater than `0`. The following items are
addressed in this PR.

1. Fix issue that causes ocean_prod tasks not to be triggered (issue
[#2725](https://github.com/NOAA-EMC/global-workflow/issues/2725)). A new
diag_table was added (called `diag_table_replay`) that is used only when
`REPLAY_ICS` is set to `YES`. This diag_table accounts for the offset
that occurs when using replay IC's.
2. Fix issue that causes atmos_prod tasks not to be triggered for the
first lead time (e.g. f003) (issue
[#2754](https://github.com/NOAA-EMC/global-workflow/issues/2754)). When
`OFFSET_START_HOUR` is greater than `0`, the first `fhr` is
`${OFFSET_START_HOUR}+(${DELTIM}/3600)`, which is defined in
`forecast_predet.sh` and will allow data for the first lead time to be
generated. The filename with this lead time will still be labelled with
`OFFSET_START_HOUR`.
3. Minor modifications were made to the extractvars task so that atmos
data from replay cases can be processed.

This PR was split from PR #2680. 

Refs #2725, #2754 

---------

Co-authored-by: Walter Kolczynski - NOAA <Walter.Kolczynski@noaa.gov>