Adjust mem if-block and fix job list for JEDI
- Incorporate reviewer feedback and adjust first[last_mem]
if-block in master_gfs.yaml.j2 to only check if RUN=enkfgdas
when setting ensemble mem values and not check for NMEM_ENS>0
- Update task list when DO_JEDIATMENS=YES to include the
stage_ic job in the hybrid_after_eupd_tasks group instead of
the hybrid_tasks. Now the enkfgdasstage_ic job shows up in
the xml, it was missing before.

Refs #2475