Correct ocean `MOM.res_#` stage copy (#2868)
This PR corrects a bug in the staging job for ocean `MOM.res_#` IC
files. The `OCNRES` value was coming in as an integer (e.g. `25`) but
the `ocean.yaml.j2` file was checking for `"025"`. Correct to now set
OCNRES to be three digits in staging script and also correct the for
loop range to include third file.

Resolves #2864