Reenable Orion Cycling Support (#2877)
This updates the model hash to include the UPP update needed to be able
to run the post processor on Orion, thus reenabling support on that
system.

A note on the UPP: it is using a newer version of g2tmpl that requires a
separate spack-stack 1.6.0 installation. This version of g2tmpl will be
standard in spack-stack 1.8.0, but for now requires loading separate
modules for the UPP.

A note on running analyses on Orion: due to a yet-unknown issue causing
the BUFR library to run much slower on Orion when compared with Rocky 8,
the GSI and GDASApp are expected to run significantly slower than on any
other platform (on the order of an hour longer).

Lastly, I made adjustments to the build_all.sh script to send more cores
to compiling the UFS and GDASApp. Under this configuration, the GSI,
UPP, UFS_Utils, and WW3 pre/post executables finish compiling before the
UFS when run with 20 cores.

Resolves #2694 
Resolves #2851 

---------

Co-authored-by: Rahul Mahajan <aerorahul@users.noreply.github.com>
Co-authored-by: Walter.Kolczynski <Walter.Kolczynski@noaa.gov>