Adjust C768 resources for Hera (#2819)
This modifies the resources for gdasfcst (everywhere) and enkfgdaseupd
(Hera only). For the fcst job, the number of write tasks is increased to
prevent out of memory errors from the inline post. For the eupd, the
number of tasks is decreased to prevent out of memory errors. The
runtime for the eupd job was just over 10 minutes.

Resolves #2506 
Resolves #2498
Resolves #2916 

---------

Co-authored-by: Walter Kolczynski - NOAA <Walter.Kolczynski@noaa.gov>