Remove need for a ice_prod dependency check script  (#2809)
# Description
The main purpose of this PR is to remove the need for an ice_prod
dependency check script `ush/check_ice_netcdf.sh`. The original purpose
of the ice_prod dependency check script is to check for special case
dependencies where `( cyc + FHMIN ) % FHOUT_ICE )) =! 0` (more details
on this issue can be found in issue #2674 ). A bugfix for these special
cases is expected to come from a PR in the ufs-weather-model.

Resolves #2721 
Refs #2721, #2674