Update for obsproc/v1.3 and new AFWA global snow filename (#2957)
Updates for the new `obsproc/v1.3` and AFWA global snow filename.
A new UFS_UTILS tag (`ops-gfsv16.3.20`) is used for the snow filename changes.

The prepobs package is also updated to `v1.2.0` but is internal to the upstream obsproc
package so only the `prepobs_run_ver` variables are updated for dev mode usage.

Additionally, workflow updates are made to allow the EMCSFC_SFC_PREP (sfcprep)
job to be run as part of the rocoto mesh in dev/emc mode. A new switch `DO_SFCPREP`
is added to `config.base.emc.dyn` and is set to "NO" by default. Setting it to "YES" 
will add the `sfcprep` job to the rocoto mesh ahead of the `prep` job and will generate
the snogrb files that are usually generated in ops and used via the GDA.

Refs #2913