Correct metp cycledefs
To avoid running metp on days where there is no gfs, the cycledefs
are adjusted somewhat. First, if the interval is >= 24h, the metp
cycledef will be identical to gfs. If the interval is < 24h, it
remains 18z every day (except the last). Second, a last_gfs is
added so metp will run on for the last gfs cycle even if it there
is no gdas cycle for 18z that day. This required computing the real
gfs end date to use as the last cycle.