Re-enable metp for forecast-only
Turns metp back on for forecast-only mode. The metp cycledef follows
the same practice of cycled, where if the interval is under 24, it
will run every 18z. But if the interval is 24+, the cycledef will
follow the gfs cycledef.

Since there is no gdasarch task to key off of when gfsarch does not
exist for the 18z cycle, the dependency is now more complicated and
progressively works backwards to find a cycle that exists, then uses
the gfsarch for that cycle.