Make aerosol output frequency variable
The `AERO_HISTORY.rc` file that controls the GOCART output is turned
into a template usint `@[ ]` syntax to allow for variable output
times. A new "parsing" script is created for GOCART that handles
parsing the `AERO_HISTORY.rc` file through `atparse`. Other manip-
ulation of these files is also moved to the new parsing script to be
consistent with other components.

A new variable, `FHOUT_AERO` is introduced to control the output
frequency of the GOCART output. This is initially set to 3 to match
other components (this is a change from the current value of 6).
Since some aerosol fields are also included as part of the atmos
output, it may be desirable to meld this back into `FHOUT` later.
However, the atmos output has two frequencies (`FHOUT_HF` and
`FHOUT`), a feature not supported by other components.

The RUN lists for aero fcst, aero analysis, and waves is simplified
to avoid confusion, since there is no need for them to only be set
if the components are on.

Resolves #2072